+++ src/primitive/mod.rs
@@ -1050,6 +1050,7 @@ impl Primitive {
             Primitive::ImageEncode => media::image_encode(env)?,
             Primitive::GifEncode => media::gif_encode(env)?,
             Primitive::AudioEncode => media::audio_encode(env)?,
+            Primitive::Project => env.dyadic_rr_env(media::project)?,
             Primitive::Layout => env.dyadic_oo_env(media::layout_text)?,
             Primitive::Fft => algorithm::fft(env)?,
             Primitive::Stringify